Move metric middleware to docs example dir.
