+++ examples/nodejs/doc-example-files/example-metric-middleware.ts
